+++ CHANGELOG.md
@@ -39,6 +39,10 @@ Section Order:
 
 - Improved explanatory text in the campaign table a bit
 
+### Removed
+
+- Unused constants
+
 ## [2.4.2] - 2025-06-16
 
 ### Fixed